Supporting header file for local search
This is a supporting header file to solve uncapacitated facility location problem using the local search moves of add, delete and swap, as provided in the book “The Design Of Approximation Algorithms” by David P. Williamson, David B. Shmoys". It is verified using the OR library test cases for correctness.